third_party/WebKit/LayoutTests/imported/wpt/dom/lists/DOMTokenList-iteration-expected.txt

This is a testharness.js-based test.
FAIL DOMTokenList iteration: keys, values, etc. assert_equals: expected function "function values() { [native code] }" but got function "function () { [native code] }"
Harness: the test ran to completion.
